Proper RAII Image handling
Image is now broken out into a class that handles its own memory. No
need to worry about forgetting to free. The shared_ptr was needed as we
don't own the pointer to the image data, and the library should be left
to free it, so it's not just a pointer into a memcopy'able buffer or
similar. (This method is actually better for performance than copying
buffers around anyway :) )